Fold the Triton kernels into the Gluon module; drop the _gluon suffix
`_triton_kernels/` holds the Triton *implementation* of an op, which a public wrapper can
select or fall back to -- `attention/pa_mqa_logits.py` picks between the Gluon kernel and the
`_triton_kernels/` one at import time. Our delta and CSR-gather kernels are not that: there is
no Triton backward to fall back to, they are the memory-bound phases of this one kernel.

Filing them under `_triton_kernels/attention/sparse_attention_dsv4_bwd.py` put them directly
next to `_triton_kernels/attention/sparse_attention_dsv4.py`, which IS the forward's Triton
path, so the pair read as a Triton fwd/bwd that does not exist. They now live beside the Gluon
kernels they serve, which is what every other Gluon module here does -- mla_gluon, pa_decode_gluon
and pa_mqa_logits all keep their plain @triton.jit helpers inline and none import from
`_triton_kernels/`.

The `_gluon` suffix goes with it: the file is already in `gluon/`, four of the seven modules
there carry no suffix, and `pa_mqa_logits` is the same basename in all three trees without one.
Same for `sparse_mla_bwd_dq_gluon` -> `sparse_mla_bwd_dq`, which its sibling never had.

No functional change. 6/6 tests pass on gfx950 (MI355X).
